Utworzenie API do pobierania listy zwierząt

+++ Cwiczenie_4/Rest_API/Data/Animal.cs
@@ -4,7 +4,7 @@ public class Animal
 {
     public int Id { get; set; }
     public string Name { get; set; }
-    public int Mass { get; set; }
+    public double Mass { get; set; }
     public string Color { get; set; }
     public Category Category { get; set; }
 }
